+++ utils/embed/deps.ts
@@ -1,4 +1,4 @@
 // @deno-types="https://npm.tfl.dev/v86/@types/react@~18.0/index.d.ts"
 export { useEffect, useState } from "https://npm.tfl.dev/react";
-export { default as jumper } from "https://tfl.dev/@truffle/utils@0.0.3/jumper/jumper.ts";
+export { default as jumper } from "../jumper/jumper.ts";
 export { useSignal } from "https://tfl.dev/@truffle/state@~0.0.5/mod.ts";
